% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregation}
\alias{aggregation}
\title{Feature Aggregation}
\usage{
aggregation(efs, metric, f = NULL)
}
\arguments{
\item{efs}{A matrix of selected features}

\item{metric}{string indicating the type of aggregation.
Avialable options are \code{"CLA"} (Complete Linear),
 \code{"EM"} (Ensemble Mean), \code{"ES"} (Ensemble Stability), and
 \code{"EE"} (Ensemble Exponential)}

\item{f}{The number of features desired.  Default \code{f = NULL}}
}
\value{
\item{agg}{Aggregated list of features}
}
\description{
Compiles matrix of ranked features via user defined 'metric'
}
\examples{
# test data
ranks <- replicate(5, sample(seq(50), 50))
row.names(ranks) <- paste0("V", seq(50))

aggregation(ranks, "CLA")
}
\author{
Charles Determan Jr
}
\references{
Abeel T., Helleputte T., Van de Peer Y., Dupont P., Saeys Y.
(2010) \emph{Robust biomarker identification for cancer diagnosis with
ensemble feature selection methods}. Bioinformatics 26(3) 392-398.

Meinshausen N., Buhlmann P. (2010) \emph{Stability selection}.
J.R. Statist. Soc. B. 72(4) 417-473.

Haury A., Gestraud P., Vert J. (2011) \emph{The Influence of Features
Selection Methods on Accuracy, Stability, and Interpretability of
Molecular Signatures}.  PLoS ONE 6(12) e28210. doi:
10.1371/journal.pone.0028210.
}
\seealso{
\code{\link{CLA}}, \code{\link{ES}}, \code{\link{EM}},
\code{\link{EE}}
}

