% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stability.R
\name{canberra_stability}
\alias{canberra_stability}
\title{Canberra Stability}
\usage{
canberra_stability(x, y)
}
\arguments{
\item{x}{numeric vector of ranks}

\item{y}{numeric vector of ranks with compatible length to x}
}
\value{
Returns the canberra stability for the two vectors
}
\description{
Calculates canberra stability between two ranked lists.  In
brief, the raw canberra distance is scaled to a [0,1] distribution by the
maximum canberra metric.  Lastly, this value is subtracted from 1 to provide
the same interpretation as the other stability metrics whereby 1 is
identical and 0 is no stability.
}
\examples{
# Canberra demo
v1 <- seq(10)
v2 <- sample(v1, 10)
canberra(v1, v2)

canberra_stability(v1, v2)
}
\author{
Charles E. Determan Jr.
}
\references{
Jurman G., Merler S., Barla A., Paoli S., Galea A., &
Furlanello C. (2008) \emph{Algebraic stability indicators for ranked lists
in molecular profiling}. Bioinformatics 24(2): 258-264.

He. Z. & Weichuan Y. (2010) \emph{Stable feature selection for biomarker
discovery}. Computational Biology and Chemistry 34 215-225.
}

