% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.generation.R
\name{create.corr.matrix}
\alias{create.corr.matrix}
\title{Correlated Multivariate Data Generator}
\usage{
create.corr.matrix(U, k = 4, min.block.size = 2, max.block.size = 5)
}
\arguments{
\item{U}{Numeric matrix}

\item{k}{Correlation Perturbation - The higher k, the more the data
is perturbed. Default \code{k = 4}}

\item{min.block.size}{minimum number of variables to correlate
Default \code{min.block.size = 2}}

\item{max.block.size}{maximum number of variables to correlate
Default \code{max.block.size = 5}}
}
\value{
Matrix of dimension \code{dim(U)} with correlations induced
between variables
}
\description{
Generates a matrix of dimensions \code{dim(U)} with
induced correlations.  Blocks of variables are randomly assigned and
correlations are induced.  A noise matrix is applied
to the final matrix to perturb 'perfect' correlations.
}
\note{
Output does not contain classes, may provide externally as
classes are irrelevant in this function.
}
\author{
Charles E. Determan Jr.
}
\seealso{
\code{\link{create.random.matrix}},
\code{\link{create.discr.matrix}}
}

