% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/denovo.grid.R.R
\name{denovo.grid}
\alias{denovo.grid}
\title{Denovo Grid Generation}
\usage{
denovo.grid(data, method, res)
}
\arguments{
\item{data}{data of method to be tuned}

\item{method}{vector indicating the models to generate grids.
Available options are \code{"plsda"} (Partial Least Squares
Discriminant Analysis),
 \code{"rf"} (Random Forest), \code{"gbm"} (Gradient Boosting Machine),
 \code{"svm"} (Support Vector Machines), \code{"glmnet"}
 (Elastic-net Generalized Linear Model),
 and \code{"pam"} (Prediction Analysis of Microarrays)}

\item{res}{Resolution of model optimization grid.}
}
\value{
A list containing dataframes of all combinations of
parameters for each model:
}
\description{
Greates grid for optimizing selected models
}
\examples{

# random test data
set.seed(123)
dat.discr <- create.discr.matrix(
    create.corr.matrix(
        create.random.matrix(nvar = 50, 
                             nsamp = 100, 
                             st.dev = 1, 
                             perturb = 0.2)),
    D = 10
)

df <- data.frame(dat.discr$discr.mat, .classes = dat.discr$classes)

# create tuning grid
denovo.grid(df, "gbm", 3)
}
\author{
Charles Determan Jr
}
\seealso{
\code{"expand.grid"} for generating grids of specific
parameters desired.  However, NOTE that you must still convert
the generated grid to a list.
}

