% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/perm.class.R
\name{extract.args}
\alias{extract.args}
\title{Argument extractor}
\usage{
extract.args(fs.model, method)
}
\arguments{
\item{fs.model}{Previously fit fs.stability model}

\item{method}{Which model to extract from}
}
\value{
args List of model arguments
}
\description{
Extract arguments from previously fs.stability models
}

