% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit.only.model.R
\name{fit.only.model}
\alias{fit.only.model}
\title{Fit Models without Feature Selection}
\usage{
fit.only.model(X, Y, method, p = 0.9, optimize = TRUE, tuning.grid = NULL,
  k.folds = if (optimize) 10 else NULL, repeats = if (optimize) 3 else NULL,
  resolution = if (optimize) 3 else NULL, metric = "Accuracy",
  allowParallel = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{X}{A scaled matrix or dataframe containing numeric values of
each feature}

\item{Y}{A factor vector containing group membership of samples}

\item{method}{A vector listing models to be fit.
Available options are \code{"plsda"} (Partial Least Squares Discriminant
Analysis), \code{"rf"} (Random Forest), \code{"gbm"} (Gradient Boosting
Machine), \code{"svm"} (Support Vector Machines), \code{"glmnet"}
(Elastic-net Generalized Linear Model), and \code{"pam"} (Prediction
Analysis of Microarrays)}

\item{p}{Percent of data to by 'trained'}

\item{optimize}{Logical argument determining if each model should be
optimized. Default \code{"optimize = TRUE"}}

\item{tuning.grid}{Optional list of grids containing parameters to optimize
for each algorithm. Default \code{"tuning.grid = NULL"} lets function
create grid determined by \code{"res"}}

\item{k.folds}{Number of folds generated during cross-validation.
Default \code{"k.folds = 10"}}

\item{repeats}{Number of times cross-validation repeated.
Default \code{"repeats = 3"}}

\item{resolution}{Resolution of model optimization grid.
Default \code{"resolution = 3"}}

\item{metric}{Criteria for model optimization.
Available options are \code{"Accuracy"} (Predication Accuracy),
\code{"Kappa"} (Kappa Statistic), and \code{"AUC-ROC"} (Area Under the
Curve - Receiver Operator Curve)}

\item{allowParallel}{Logical argument dictating if parallel processing
is allowed via foreach package.
Default \code{allowParallel = FALSE}}

\item{verbose}{Logical argument if should output progress}

\item{...}{Extra arguments that the user would like to apply to the models}
}
\value{
\item{Methods}{Vector of models fit to data}

\item{performance}{Performance metrics of each model and
bootstrap iteration}

\item{specs}{List with the
following elements:}

\itemize{
 \item{total.samples: Number of samples in original dataset}
 \item{number.features: Number of features in orginal dataset}
 \item{number.groups: Number of groups}
 \item{group.levels: The specific levels of the groups}
 \item{number.observations.group: Number of observations in each group}}
}
\description{
Applies models to high-dimensional data for classification.
}
\examples{
dat.discr <- create.discr.matrix(
    create.corr.matrix(
        create.random.matrix(nvar = 50, 
                             nsamp = 100, 
                             st.dev = 1, 
                             perturb = 0.2)),
    D = 10
)$discr.mat


vars <- dat.discr[,1:(ncol(dat.discr)-1)]
groups <- as.factor(dat.discr[,ncol(dat.discr)])

fit <- fit.only.model(X=vars, 
                      Y=groups, 
                      method="plsda", 
                      p = 0.9)
}
\author{
Charles Determan Jr
}

