% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelList.R
\name{modelList}
\alias{modelList}
\title{Model List}
\usage{
modelList()
}
\value{
A data.frame containing:
\item{methods}{The abbreviated code for the method}
\item{description}{Full name of the method}
}
\description{
Provide a list of currently implemented methods 
for OmicsMarkeR.
}
\examples{
modelList()
}
\author{
Charles Determan Jr.
}

