% Generated by roxygen2 (4.0.2): do not edit by hand
\name{performance.stats}
\alias{performance.stats}
\title{Performance Statistics (Internal for \code{perf.calc})}
\usage{
performance.stats(pred, obs)
}
\arguments{
\item{pred}{vector of groups predicted by a fitted classification model}

\item{obs}{vector of groups from the original dataset}
}
\value{
Returns confusion matrix and ROC performance statistics including
Accuracy, Kappa, ROC.AUC, Sensitivity, Specificity, Positive Predictive Value, and Negative Predictive Value
}
\description{
Calculates confusion matrix and ROC statistics comparing the results of the fitted models
to the observed groups.
}
\seealso{
caret function \code{\link{confusionMatrix}}
}

