% Generated by roxygen2 (4.0.0): do not edit by hand
\name{predicting}
\alias{predicting}
\title{Model Group Prediction}
\usage{
predicting(method, modelFit, orig.data, indicies, newdata, param = NULL)
}
\arguments{
\item{method}{String of the model to be evaluated}

\item{modelFit}{The fitted model being evaluated}

\item{orig.data}{The orginal data before subsetting training sets.  Required to have the 'observed' group membership}

\item{indicies}{The indicies for the training subsets}

\item{newdata}{The testing data to predict group membership}

\item{param}{The parameters being fit to the model (Determined by model optimization).}
}
\value{
Returns a list of predicted group membership
}
\description{
This function evaluates a single fitted model and returns the predicted group memberships.
}

