% Generated by roxygen2 (4.0.2): do not edit by hand
\name{svm.weights}
\alias{svm.weights}
\title{SVM Multiclass Weights Ranking}
\usage{
\method{svm}{weights}(model)
}
\arguments{
\item{model}{A fitted SVM model of multiclass}
}
\value{
Vector of feature weights
}
\description{
This calculates feature weights for multiclass Support Vector Machine (SVM) problems
}
\references{
Guyon I. et. al. (2010) \emph{Gene Selection for Cancer Classification using Support Vector Machines}.
Machine Learning 46 389-422.
}

