% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/training.R
\name{training}
\alias{training}
\title{Model Training}
\usage{
training(data, method, tuneValue, obsLevels, theDots = NULL)
}
\arguments{
\item{data}{Dataframe consisting of both numeric feature values and a single column
named '.classes' to denoted group membership.}

\item{method}{String dictating which model to fit}

\item{tuneValue}{List of parameters to be applied to the specific model}

\item{obsLevels}{Observed group levels}

\item{theDots}{List of additional parameters to be applied to the specific model}
}
\value{
\item{fit}{Fitted model with list with the following elements:}

\itemize{
 \item{xNames: Names of the features}
 \item{tuneValue: Parameters applied to the fitted model}
 \item{obsLevels: Observed levels of the groups}}
}
\description{
This fits each model with the defined parameters
}
\author{
Charles Determan Jr
}

