% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.generation.R
\name{create.random.matrix}
\alias{create.random.matrix}
\title{Random Multivariate Data Generator}
\usage{
create.random.matrix(nvar, nsamp, st.dev = 1, perturb = 0.2)
}
\arguments{
\item{nvar}{Number of features (i.e. variables)}

\item{nsamp}{Number of samples}

\item{st.dev}{The variation (i.e. standard deviation) that is typical
in datasets of interest to the user.  Default \code{spread = 1}}

\item{perturb}{The amount of perturbation to the normal distribution.
Default \code{perturb = 0.2}}
}
\value{
Matrix of dimension \code{nvar} by \code{nsamp}
}
\description{
Generates a matrix of dimensions \code{nvar} by
\code{nsamp} consisting of random numbers generated from a normal
distriubtion.  This normal distribution is then perturbed to more
accurately reflect experimentally acquired multivariate data.
}
\examples{
# Create Multivariate Matrices

# Random Multivariate Matrix

# 50 variables, 100 samples, 1 standard devation, 0.2 noise factor

rand.mat <- create.random.matrix(nvar = 50, 
                                 nsamp = 100, 
                                 st.dev = 1, 
                                 perturb = 0.2)


# Induce correlations in a numeric matrix

# Default settings
# minimum and maximum block sizes (min.block.size = 2, max.block.size = 5)
# default correlation purturbation (k=4)
# see ?create.corr.matrix for citation for methods

corr.mat <- create.corr.matrix(rand.mat)


# Induce Discriminatory Variables

# 10 discriminatory variables (D = 10)
# default discrimination level (l = 1.5)
# default number of groups (num.groups=2)
# default correlation purturbation (k = 4)

dat.discr <- create.discr.matrix(corr.mat, D=10)
}
\author{
Charles E. Determan Jr.
}
\references{
Wongravee, K., Lloyd, G R., Hall, J., Holmboe, M. E., &
Schaefer, M. L. (2009). \emph{Monte-Carlo methods for determining optimal
number of significant variables. Application to mouse urinary profiles.}
Metabolomics, 5(4), 387-406. http://dx.doi.org/10.1007/s11306-009-0164-4
}
\seealso{
\code{\link{create.corr.matrix}}, \code{\link{create.discr.matrix}}
}

