% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feature.table.R
\name{feature.table}
\alias{feature.table}
\title{Feature Consistency Table}
\usage{
feature.table(features, method)
}
\arguments{
\item{features}{A fs.stability fitted object}

\item{method}{Algorithm of interest
Available options are \code{"plsda"} (Partial Least Squares
Discriminant Analysis), \code{"rf"} (Random Forest),
\code{"gbm"} (Gradient Boosting Machine), \code{"svm"}
(Support Vector Machines), \code{"glmnet"}
(Elastic-net Generalized Linear Model), and \code{"pam"}
(Prediction Analysis of Microarrays)}
}
\value{
A data frame containing:
 \item{features}{Features identified by model}
 \item{consistency}{Number of iterations feature was identified}
 \item{frequency}{Frequency of iterations the feature was identified}
}
\description{
Extracts and sorts the features identified for a given method.
}
\examples{
dat.discr <- create.discr.matrix(
    create.corr.matrix(
        create.random.matrix(nvar = 50, 
                             nsamp = 100, 
                             st.dev = 1, 
                             perturb = 0.2)),
    D = 10
)

vars <- dat.discr$discr.mat
groups <- dat.discr$classes

fits <- fs.stability(vars, 
                     groups, 
                     method = c("plsda", "rf"), 
                     f = 10, 
                     k = 3, 
                     k.folds = 10, 
                     verbose = 'none')

feature.table(fits, "plsda")
}
\author{
Charles Determan Jr
}

