% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/performance.metrics.R
\name{performance.metrics}
\alias{performance.metrics}
\title{Performance Metrics of fs.stability or fs.ensembl.stability object}
\usage{
performance.metrics(fit.model, digits = max(3, getOption("digits") - 3))
}
\arguments{
\item{fit.model}{An fs.stability or fs.ensembl.stability object}

\item{digits}{How many digits to round values}
}
\value{
Dataframe of performance statistics by model
}
\description{
This will provide a concise data.frame of confusion matrix
and ROC statistics from the results of \code{fs.stability} or
\code{fs.ensembl.stability}.
}
\examples{
dat.discr <- create.discr.matrix(
    create.corr.matrix(
        create.random.matrix(nvar = 50, 
                             nsamp = 100, 
                             st.dev = 1, 
                             perturb = 0.2)),
    D = 10
)

vars <- dat.discr$discr.mat
groups <- dat.discr$classes

fits <- fs.stability(vars, 
                     groups, 
                     method = c("plsda", "rf"), 
                     f = 10, 
                     k = 3, 
                     k.folds = 10, 
                     verbose = 'none')

performance.metrics(fits)
}
\author{
Charles E. Determan Jr.
}

