% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{spearman}
\alias{spearman}
\title{Spearman Rank Correlation Coefficient}
\usage{
spearman(x, y)
}
\arguments{
\item{x}{numeric vector of ranks}

\item{y}{numeric vector of ranks with compatible length to x}
}
\value{
Returns the spearman rank coefficient for the two vectors
}
\description{
Calculates spearman rank correlation between two vectors
}
\examples{
# Spearman demo
v1 <- seq(10)
v2 <- sample(v1, 10)
spearman(v1, v2)
}

