\name{svm.weights}
\alias{svm.weights}
\title{SVM Multiclass Weights Ranking}
\usage{
  svm.weights(model)
}
\arguments{
  \item{model}{A fitted SVM model of multiclass}
}
\value{
  Vector of feature weights
}
\description{
  This calculates feature weights for multiclass Support
  Vector Machine (SVM) problems
}
\references{
  Guyon, I., Weston, J., Barnhill, S. & Vapnik, V. (2002)
  \emph{Gene Selection for Cancer Classification using
  Support Vector Machines}. Machine Learning 46, 389–422.
}

